requirements.txt
numpy~=1.23.1
torch~=1.8.1+cu101
scikit-learn~=0.24.2
tqdm~=4.64.0
faiss-gpu~=1.6.1
